+++ packages/cli/api-importers/openapi/openapi-ir-parser/src/openapi/v3/extensions/getSecuritySchemeNameAndEnvvars.ts
@@ -7,10 +7,12 @@ export interface HeaderSecuritySchemeNames {
     name?: string;
     env?: string;
     prefix?: string;
+    placeholder?: string;
 }
 export interface SecuritySchemeNames {
     name?: string;
     env?: string;
+    placeholder?: string;
 }
 export interface BasicSecuritySchemeNames {
     username?: SecuritySchemeNames;