+++ seed/swift-sdk/basic-auth-pw-omitted/Tests/Core/ClientErrorTests.swift
@@ -16,7 +16,7 @@ import Testing
         let client = BasicAuthPwOmittedClient(
             baseURL: "https://api.fern.com",
             username: "<username>",
-            password: "<password>",
+            password: "",
             urlSession: stub.urlSession
         )
 
@@ -48,7 +48,7 @@ import Testing
         let client = BasicAuthPwOmittedClient(
             baseURL: "https://api.fern.com",
             username: "<username>",
-            password: "<password>",
+            password: "",
             urlSession: stub.urlSession
         )
 
@@ -80,7 +80,7 @@ import Testing
         let client = BasicAuthPwOmittedClient(
             baseURL: "https://api.fern.com",
             username: "<username>",
-            password: "<password>",
+            password: "",
             urlSession: stub.urlSession
         )
 
@@ -114,7 +114,7 @@ import Testing
         let client = BasicAuthPwOmittedClient(
             baseURL: "https://api.fern.com",
             username: "<username>",
-            password: "<password>",
+            password: "",
             urlSession: stub.urlSession
         )
 
@@ -146,7 +146,7 @@ import Testing
         let client = BasicAuthPwOmittedClient(
             baseURL: "https://api.fern.com",
             username: "<username>",
-            password: "<password>",
+            password: "",
             urlSession: stub.urlSession
         )
 
@@ -180,7 +180,7 @@ import Testing
         let client = BasicAuthPwOmittedClient(
             baseURL: "https://api.fern.com",
             username: "<username>",
-            password: "<password>",
+            password: "",
             urlSession: stub.urlSession
         )
 
@@ -212,7 +212,7 @@ import Testing
         let client = BasicAuthPwOmittedClient(
             baseURL: "https://api.fern.com",
             username: "<username>",
-            password: "<password>",
+            password: "",
             urlSession: stub.urlSession
         )